+++ src/pages/Cources/Filters/CategoriesField/styles.module.scss
@@ -1,0 +1,23 @@
+.categories {
+  display: flex;
+  gap: 20px;
+
+  label {
+    cursor: pointer;
+
+    input[type=radio] {
+      display: none;
+    }
+
+    &:has(input[type=radio]:checked) {
+      background: var(--orange-50);
+      transition: var(--hover);
+    }
+  }
+}
+
+.category {
+  padding: 10px;
+  border-radius: 5px;
+  transition: var(--hover);
+}
