+++ src/pages/Cources/Filters/LevelField/styles.module.scss
@@ -1,0 +1,7 @@
+.level {
+  margin-left: auto;
+  background: var(--black-15);
+  color: white;
+  padding: 10px;
+  border-radius: 5px;
+}
