+++ src/pages/Cources/Filters/styles.module.scss
@@ -1,0 +1,5 @@
+.filters {
+  display: flex;
+  align-items: center;
+  gap: 35px;
+}
