+++ src/pages/Cources/styles.module.scss
@@ -27,63 +27,3 @@
   border-radius: 10px;
   background: var(--black-15);
 }
-
-.filters {
-  display: flex;
-  align-items: center;
-  gap: 35px;
-}
-
-.categories {
-  display: flex;
-  gap: 20px;
-
-  label {
-    cursor: pointer;
-
-    input[type=radio] {
-      display: none;
-    }
-
-    &:has(input[type=radio]:checked) {
-      background: var(--orange-50);
-      transition: var(--hover);
-    }
-  }
-}
-
-.category {
-  padding: 10px;
-  border-radius: 5px;
-  transition: var(--hover);
-}
-
-.level {
-  margin-left: auto;
-  background: var(--black-15);
-  color: white;
-  padding: 10px;
-  border-radius: 5px;
-}
-
-.search {
-  display: flex;
-  width: 300px;
-}
-
-.searchInput {
-  width: 100%;
-  height: 42px;
-  padding-left: 10px;
-  border-radius: 5px 0 0 5px;
-  background: var(--black-15);
-  color: white;
-}
-
-.searchButton {
-  width: 50px;
-  height: 42px;
-  background: var(--black-20);
-  border-radius: 0 5px 5px 0;
-  cursor: pointer;
-}
